refactor(apply_fixes): More efficient testing of CLI
The integration tests are a somewhat complicated setup and not very fast.
Thus, we prefer testing the trivial case of asserts on CLI args via unit
tests.
