+++ dwyu/apply_fixes/cli.py
@@ -1,0 +1,143 @@
+import logging
+import sys
+from argparse import ArgumentParser, Namespace, RawDescriptionHelpFormatter
+from pathlib import Path
+
+log = logging.getLogger()
+
+
+def cli() -> Namespace:
+    parser = ArgumentParser(
+        formatter_class=RawDescriptionHelpFormatter,
+        description="""
+This script expects the user has invoked DWYU in the given workspace beforehand and by doing so generated DYWU report files in the output path.
+Beware, running this script multiple times with the same report files might raise a lot of warnings due to the fixed code no longer matching the old reports.
+
+This script has to deduce the 'bazel-bin' dir in the Bazel output base by executing the 'bazel info' command.
+Have a look at the '--bazel-args' and '--bazel-startup-args' options, to ensure this script uses the same Bazel configuration as was used to execute DWYU.
+
+Beware that there are limits on the fixes buildozer can execute.
+Buildozer works on a limited set of information, since it looks at BUILD files before macro and alias expansion:
+- If targets are generated by macros buildozer cannot edit them.
+- The DWYU aspect runs after alias expansion and thus reports the actual names instead of the alias names.
+  However, only the alias name is visible to buildozer in a BUILD file in the dependencies of a target.
+
+The script expects 'bazel' to be available on PATH.
+    """.strip(),
+    )
+    parser.add_argument(
+        "--fix-unused-deps",
+        action="store_true",
+        help="Automatically remove unused dependencies.",
+    )
+    parser.add_argument(
+        "--fix-deps-which-should-be-private",
+        action="store_true",
+        help="Automatically move 'deps' to 'implementation_deps'.",
+    )
+    parser.add_argument(
+        "--fix-missing-deps",
+        action="store_true",
+        help="""
+        Automatically search and add dependencies providing headers for which a direct dependency is missing.
+        This is based on a heuristic and thus is not guaranteed to work.
+        """,
+    )
+    parser.add_argument(
+        "--fix-all",
+        action="store_true",
+        help="Perform all available automatic fixes.",
+    )
+    parser.add_argument(
+        "--dry-run",
+        action="store_true",
+        help="Don't apply fixes. Report the buildozer commands and print the adapted BUILD files to stdout.",
+    )
+    parser.add_argument(
+        "--workspace",
+        metavar="PATH",
+        type=Path,
+        help="""
+        Workspace for which DWYU reports are gathered and fixes are applied to the source code.
+        If no dedicated workspace is provided, we assume we are running from within the workspace for which the DWYU reports have been generated and determine the workspace root automatically.
+        """,
+    )
+    parser.add_argument(
+        "--reports-search-path",
+        metavar="PATH",
+        type=Path,
+        help="""
+        Search recursively for DWYU report files below this path.
+        This disables automatically deducing the 'bazel-bin' dir.
+        When used together with the '--dwyu-log-file' option, please note the extra constraints documented in the '--dwyu-log-file' option.
+        """,
+    )
+    parser.add_argument(
+        "--dwyu-log-file",
+        metavar="PATH",
+        type=Path,
+        help="""
+        Extract the relevant DWYU report files from a file containing the command line output of executing DWYU.
+        This is the recommended approach, since it is the most efficient one for large workspaces and it cannot find outdated DWYU report files by chance.
+        When you combine this option with '--reports-search-path', beware that '--reports-search-path' has to point exactly to the 'bazel-bin' directory and cannot point to a sub directory.
+        """,
+    )
+    parser.add_argument(
+        "--use-cquery",
+        action="store_true",
+        help="""
+        This script uses by default 'bazel query' to find missing dependencies.
+        Your project might use select statements to exchange dependencies.
+        In such cases you should use 'bazel cquery' to allow this script understanding the dependency tree properly.
+        To ensure the cquery command uses the same configuration as your DWYU execution, use the options '--bazel-args' and '--bazel-startup-args'.
+        """,
+    )
+    parser.add_argument(
+        "--bazel-args",
+        type=str,
+        metavar="STRING",
+        help="""
+        This script executes bazel commands.
+        If those need to be configured to be compatible to your workspace and use the same configuration as was used to generate the DWYU reports, you can provide bazel arguments here.
+        Arguments have to be provided as continuous string, e.g.: --bazel-args='--foo --tick=tock'.
+        """,
+    )
+    parser.add_argument(
+        "--bazel-startup-args",
+        type=str,
+        metavar="STRING",
+        help="""
+        This script executes bazel commands.
+        If those need startup arguments to be compatible to your workspace and use the same configuration as was used to generate the DWYU reports, you can provide bazel startup arguments here.
+        Arguments have to be provided as continuous string, e.g.: --bazel-startup-args='--foo --tick=tock'.
+        """,
+    )
+    parser.add_argument(
+        "--buildozer",
+        metavar="PATH",
+        type=Path,
+        help="Use a custom buildozer binary instead of the bundled one. Expects an absolute path.",
+    )
+    parser.add_argument(
+        "--buildozer-args",
+        type=str,
+        metavar="STRING",
+        help="""
+        Forward arguments to buildozer.
+        Arguments have to be provided as continuous string, e.g.: --buildozer-args='-foo -tick=tock'.
+        """,
+    )
+    parser.add_argument(
+        "--verbose",
+        action="store_true",
+        help="Announce intermediate steps.",
+    )
+
+    args = parser.parse_args()
+
+    has_explicit_fix_option = any((args.fix_unused_deps, args.fix_deps_which_should_be_private, args.fix_missing_deps))
+    if not has_explicit_fix_option and not args.fix_all:
+        logging.fatal("Please choose at least one of the 'fix-..' options")
+        sys.exit(1)
+
+    return args
